requirements.txt
absl-py==0.9.0
alabaster==0.7.12
argh==0.26.2
asn1crypto==1.2.0
astor==0.8.1
astroid==2.3.3
atomicwrites==1.3.0
attrs==19.3.0
Automat==0.8.0
autopep8==1.4.4
Babel==2.8.0
backcall==0.1.0
bcrypt==3.1.7
bleach==1.5.0
cachetools==4.0.0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
cloudpickle==1.2.2
colorama==0.4.3
constantly==15.1.0
cryptography==2.8
cssselect==1.1.0
cycler==0.10.0
decorator==4.4.1
defusedxml==0.6.0
diff-match-patch==20181111
dnspython==1.16.0
docutils==0.15.2
entrypoints==0.3
flake8==3.7.9
future==0.18.2
gast==0.2.2
google-auth==1.10.1
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.26.0
h5py==2.10.0
html5lib==0.9999999
hyperlink==19.0.0
idna==2.8
imageio==2.6.1
imagesize==1.2.0
importlib-metadata==1.3.0
incremental==17.5.0
intervaltree==3.0.2
ipykernel==5.1.3
ipython==7.11.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
jedi==0.14.1
Jinja2==2.10.3
joblib==0.14.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.1.0
jupyter-core==4.6.1
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==20.0.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
lxml==4.4.2
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.2
mccabe==0.6.1
mistune==0.8.4
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
more-itertools==8.0.2
mysql-connector-python==8.0.19
nbconvert==5.6.1
nbformat==4.4.0
networkx==2.4
notebook==6.0.2
numpy==1.17.4
numpydoc==0.9.2
oauthlib==3.1.0
olefile==0.46
opt-einsum==3.1.0
packaging==20.0
pandas==0.25.3
pandocfilters==1.4.2
paramiko==2.6.0
parsel==1.5.2
parso==0.5.2
pathtools==0.1.2
pexpect==4.7.0
pickleshare==0.7.5
Pillow==7.0.0
pluggy==0.13.1
prometheus-client==0.7.1
prompt-toolkit==3.0.2
protobuf==3.11.2
psutil==5.6.7
py==1.8.1
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycodestyle==2.5.0
pycparser==2.19
PyDispatcher==2.0.5
pydocstyle==4.0.1
pyflakes==2.1.1
Pygments==2.5.2
PyHamcrest==1.9.0
pylint==2.4.4
PyNaCl==1.3.0
pyOpenSSL==19.1.0
pyparsing==2.4.6
pyrsistent==0.15.6
PySocks==1.7.1
pytest==5.3.2
pytest-runner==5.2
python-dateutil==2.8.1
python-jsonrpc-server==0.3.2
python-language-server==0.31.2
pytz==2019.3
PyWavelets==1.1.1
pywin32==227
pywin32-ctypes==0.2.0
pywinpty==0.5.7
PyYAML==5.2
pyzmq==18.1.0
QDarkStyle==2.7
QtAwesome==0.6.0
qtconsole==4.6.0
QtPy==1.9.0
queuelib==1.5.0
requests==2.22.0
requests-oauthlib==1.3.0
rope==0.14.0
rsa==4.0
Rtree==0.8.3
scikit-image==0.16.2
scikit-learn==0.22.1
scipy==1.4.1
Scrapy==1.6.0
selenium==3.141.0
Send2Trash==1.5.0
service-identity==18.1.0
six==1.13.0
sklearn==0.0
snowballstemmer==2.0.0
sortedcontainers==2.1.0
Sphinx==2.3.1
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
spyder==4.0.1
spyder-kernels==1.8.1
tensorboard==2.1.0
tensorflow-gpu==2.1.0
tensorflow-gpu-estimator==2.1.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
tornado==6.0.3
traitlets==4.3.3
Twisted==19.10.0
typed-ast==1.4.0
ujson==1.35
urllib3==1.25.7
w3lib==1.21.0
watchdog==0.9.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.0
widgetsnbextension==3.5.1
win-inet-pton==1.1.0
wincertstore==0.2
wrapt==1.11.2
yapf==0.28.0
zipp==0.6.0
zope.interface==4.7.1
